Add keyword search functionality & refactor search. Fixes #25
To follow the Rails "fat model, skinny controller" convention,
this commit takes all the search and view logic (for the results text) out of the controller and into the model.

Scopes were added for the location and keyword queries, including
a default scope that sorts results alphabetically. The keyword query
uses regex to find a match in either the organization's name or its keywords. The regex for the name looks for an entire word that matches the query string. For the keyword, it tries to find any occurrence of that string. For example, a search for "farmer" will return results where the "farmers' market" keyword exists. However,  a search for "belm" will not return any organizations that have "Belmont" in the name, assuming "belm" is not a string present in any of the keywords.

The search results copy has been updated, and the integration
tests have been modified accordingly.