Add market_match? method to use in views
This method returns "true" is a farmers' market participates in the Market Match program. In the view, we can use this method to check and display a Market Match logo for example.

I also added in the view an example way of displaying the types of payments accepted by a farmers' market.