Add return to search results feature. Fixes #40
I also fixed another bug where going to /organizations caused the distance code (that displays the distance to the search address from the results) to throw an error because @location is blank.
I also fixed the integration test to catch this in the future by making sure the test DB has at least 1 record.