Add more robust address validations. Fixes #52
The address field will reject any entries that begin with any 5 digits and are followed by a dash.

If the entry only contains numbers and if the string is not exactly 5 characters long, we display an error message.

If the entry is exactly 5 digits, we use the "area" gem to determine whether the zip code exists. If it doesn't, we display an error message.

If the entry is blank, the /organizations index is displayed with all organizations in the database (which should be paginated in the future)

If the entry contains any string that will cause an application level Mongoid error, that error is intercepted, the user is redirected to the home page, and a friendly error message is displayed.