Remove trailing whitespace from library phone number
The phone number validation in the app does not allow creating an organization in the DB if the phone contains any characters after the last digit. While loading the data on heroku, the validation caught that some library phone numbers were not properly formatted.

I edited the scraper script to remove any whitespaces from the number. Now, the dataset is clean and can be loaded in the DB.