Configure Geocoder to use Mapquest, and cache results in Redis
We were hitting the 2500 requests/day limit with the Google Maps API, so we're switching to Mapquest, which doesn't have any limits. I've also configured Geocoder to cache results with Redis, which improves performance considerably.

The README has been updated with Redis installation instructions.

Unrelated changes:
I added more fields to the model in anticipation of the CIP data.