Add Redis config to geocoder gem so it can use REDIS constant
The problem with Redis trying to connect to localhost on Heroku was because the geocoder cache config was pointing to Redis.new, which means localhost. By using the REDIS constant, it will point either to the production URL in production, or to localhost in the dev environment.

Thanks to @ahhrrr for pointing me in the right direction!