Update organization model to match CIP data
Updated the tests to describe the changes we want to make.

Some organizations have multiple emails, URLs, faxes, and TTYs, so the field type has been changed to Array.

Some organizations don't have a constant physical location, so I removed the requirement that an address be present.

Because emails and URLs are now arrays, we can't use the validates_formatting_of gem, so I created Class methods to validate each email and URL in the array using the same regex from the validates_formatting_of gem.

I moved the phone formatting method to a helper and modified it to accept a phone number as a parameter. I also changed its name to reflect what it does.

I updated the "show" view to match the model changes.

I updated the farmers' markets json to put the url in an "urls" array.