Skip invalid records and log errors to file when populating DB
The geocoder gem automatically geocodes new records after they are validated. This takes a long time however, so I've disabled geocoding for now. Once we upload the cip data, the geocoding should be turned back on for when we allow addition of new records or updating existing ones on a limited scale.