Update rake task for local deployment
Since the main data will be coming from the API, we will no longer need a rake task to load the CIP data. Contributors who want to deploy this app locally — while we refactor the app to start consuming the API — can do so with a subset of the data, namely the libraries and farmers' markets.