Make sure street_address is present before superscripting ordinals
If street_address is not present, the superscript helper throws an error because you can't call the "split" method on nil:NilClass. I added a Factory for an organization without an address and added a test to make sure having such an organization in the search results doesn't throw an error. I also added a presence check for street_address before callling the superscript_ordinal method.