Fixes mapbox ajax display issue
Mapbox was not displaying on ajax calls because the map variable was
intialized but referring to a node that no longer existed. This
provides a try/catch hack to reinitialize the map variable if it has
not already happened. Also, this commit removes unused config files
that were causing a build error.