Fixes error in detail view when no parameters are present
When a view is accessed directly and no parameters are present the view
was throwing an error because the organization was not necessarily
found in the search results because only the first 30 results are
returned. This commit performs a second query for the specific
organization if the first query didn't include it.