Fixes invalid search test error
Invalid search test was failing because the search didn't include a
path and the orgs variable was being checked for being nil, not empty,
as it should have been. This was causing an error because the service
categories were actually not showing up on the 'no results' search
results page as they should.