Fixes no results map issue
When map had no results it was trying to fit to the bounds of the
markers. Since there aren't any when there are no results, the map was
throwing a JavaScript error. This fixes the issue by first checking
that markers exist before performing a fit to bounds operation on the
map.