Fixes #193 - masks popup arrow
The arrow on the popup is made by rotating a box 90 degrees. The
consequence of this is that the lower corner of the box overlapped on
the text when the popup was scrolled. This crops the arrow so that it
forms an actual triangle and doesn't overlay on top of the text when
scrolling.