Fixes layout issue with small result sets
Small number of results (e.g. one result) caused the footer to overlap
the search sidebar. This commit fixes that issue by adding a clear
float to the footer as well as handling the border between the search
sidebar and content when the search results occupy a variety of heights.