Test updates, bug fixes, details page additions
- List entries were using JS to handle having links inside of links.
This commit formats the list entries to not depend on JS for the full
entry link handling and instead uses CSS to position the list item
entry link in front of all content and then place embedded links on top
of that.
- Updates tests and sets those tests that refer to content that is not
displayed in the view to pending.
- Moves URLs to a partial template so the same template can be rendered
in the services as well as in the location.
- Updates responsive styles to move details page sidebar to full width
under 1024px displays.