Fix location details specs
After visit_details, we need to check for a unique element on the details page, such as the email address. #floating-results-header is present on both the results and details page, and it throws things off. Once we're sure we're on the details page, we click the first link within the #results-container, which is the Result list button.