Refactor view helpers
Move all view helpers to application_helper. They don't belong in the model. Methods defined in application_helper are available to all views, so the datalist_initializer in application_controller is no longer needed.

What used to be called "terminology" is now also handled in the application_helper via the "dynamic partial" method and has been removed from the controller.

The app should never make more requests than necessary, so the logic to only use the first part of a term that includes a slash is now in the view instead of the controller, and the app will only make one request now.

Renamed "name and "sub" to "parent" and "child" to make the code more readable.

Refactored the category partial to use a partial since it was repeating the same code twice.