Fixes #249 - fixes incorrect value in search results summary
Last page of the search results summary was displaying the wrong value
if the last page contained less entries than the per page value of the
API (currently 30). This commit caps the entry range to the total entry
count so that it doesn't flow over on the last page. Also updates a
test to check for this.