Use GA script from docs instead of analytics portal
Google's docs are not in sync with the actual portal. The tracking code snippet from the portal doesn't define "_gaq", so tracking hasn't been working since "_gaq" was not defined. I just noticed the error in the console.

I switched to the code provided in the docs, which does define "_gaq", so hopefully we'll start getting some tracking data now.