Temporarily remove the nearby feature
Fetching nearby places is the most time-consuming activity in the app.
The API method needs to be optimized, but the app should not be
automatically fetching them every time you visit the details page.
It should only fetch them if someone asks for them on the details page,
and we should add a Google Analytics event to track how many times
people click "Show nearby services".