Set canonical URL for details page.
Since this app includes various parameters in the URL when linking to a
location's details page, we can end up with many URLs that display the
same content. To gain more control over which URL appears in Google search
results, we can use the <link> element with the "rel=canonical" attribute in the <head> section of the page.