Adds manager for updating URL when Google Translate widget is changed
Updating the Google Translate widget didn't update the URL 'translate'
parameter, so the language links at the top of the page would not
correspond to the dropdown menu at the bottom of the page. This updates
the URL 'translate' parameter to the language selected in the Google
widget and reloads the page to apply the style updates the links at the
top of the page.