Fixes style issue on Firefox
There is a bug/feature of Firefox where elements with their CSS display
attribute set to a table display are invisible to absolute positioned
elements inside them. This commit adds a wrapper div around the
absolutely positioned radio button graphics, so that they don't stack
up on themselves in Firefox.