Updates result map manager
No results map warning was showing in the JS console whenever the
results map container was absent, even when no results were found. This
warning is only really of interest when there are results found, but no
map is present (indicating entries that aren't geocoded). This commit
updates the result map manager script to check whether the page is in a
no results state before checking for the map container.