Adds filter check for when add input has a value
Sometimes on page load the add input will have a value, such as when
geolocation is used, but the filters manager didn't register the add
input as showing, which led to unexpected behavior. This commit adds
logic to register the opening of the input if the add input has a value
other than an empty string at page load.