Update app to work with v2.0.0 of ohana-api
This app now points to the new demo API, which only contains a small sample data set: http://ohana-api-demo.herokuapp.com

In order to let ohana-web-search test pagination, the demo API displays one result per page.

The other notable change is replacing poltergeist with capybara-webkit, which is so much nicer. Not a single error.
I wish I had thought of switching before spending hours fighting with poltergeist. Haven't tried it on travis yet, but we'll see as soon as I push this commit.

I also got rid of the Kind search filter since that's specific to SMC, but we still need to clean up the other places where Kind is used. I also defaulted the service_areas param to an empty string instead of SMC, for obvious reasons.