Use ohana-api-test for specs
This PR basically just updates the specs to point to ohana-api-test

I created a new instance of the API on Heroku to use with specs so that we can let others play with ohana-api-demo without affecting the tests. Also, ohana-api-test defaults to 1 result per page, but I want to change ohana-api-demo to use 30 by default. This PR needs to merged first though.