Add geolocation form back to _aside partial
I figured out the original issue that caused me to add an id to the hidden field. It was because I changed the name (and therefore the id too) of the search fields from "#{css_field}-option-input" to the field name, resulting in the search fields having the same id as the hidden field. But adding an id to the hidden field was not the right solution because the location-manager is looking for an id=location, of which there are now many.

So, I reverted the hidden field back to have both name and id set to the field name, and added the "#{css_field}-option-input" id back to the search fields. Now everything is working like it should.