Temporarily remove GA manager to push to staging
We removed the GA manager from app-init a few days ago to upgrade to Universal Analytics, and at that time, the GA manager was not called in routes/about/index.js. Now that we've pulled in changes from upstream, it is, but since we removed it, it's causing asset compilation to fail.

So, for now, we're removing the reference to it until we merge PR #450 from upstream.