Moves Google Maps plugins to be inline and Code Cleanup
The plugins work fine locally using nested require() calls. However, when
using the requirejs optimizer, the plugins were concatenated in the main
application script, whereas Google Maps API is loaded asynchronously. This
led to errors in production because the local concatenated plugins were
executing before their dependencies loaded. Having the plugins inline isn't
ideal, but based on research it appears to be the approach with the least
overhead and disruption of the existing site structure.

It allows:
- Guaranteed execution of the plugins after the Google Maps APIs have loaded.
- Exposure of the module initialization method to the outside. With a nested
require() call inside the module, the init method was not accessible outside
the module scope. As inline plugins the init method is accessible again and
can be used to specify where and when the module is initialized.
- Removal of vendor scripts and the associated path settings in the
requirejs.yml config file.

Alternatives to this setup that could be investigated in the future:
(a) The Promises based approach shown here
http://stackoverflow.com/questions/12648598/amd-version-of-google-maps-v3-for-use-with-require-js/
The consequences of this approach are the introduction of a jQuery dependency
for the loading and possible non-support for Promises in IE being an issue.

(b) Include directive. See
http://stackoverflow.com/questions/19337293/requirejs-optimiser-include-nested-dependencies-for-certain-files-and-folders
The plugins could still be nested like how they were set up and worked locally
and then explicitly included in the build. However, this may still cause them
to load before the maps, so that would need to be investigated on a staging
server.

Other cleanup and adjustments:

- Adds findNestedDependencies flag to requirejs config. Nested require
calls were not being included in the optimizer.

- Removes erroneous reference to term-popup-manager which is removed.

- Makes bitmask return boolean for isOn method, instead of actual bit
numerical value.

- General code cleanup and documentation added.