Simplify search filters
This replaces all the complicated view, JS, and controller code with 2
simple text fields for searching by location and agency name. This is a
familiar interface and keep things simple and maintainable. Anything
that would be built on top of this simple foundation should be based on
user feedback.

Note that the text fields are not styled in this commit.