Improve geolocation wait text implementation
In the previous implementation, changing the text in the locale would
not update the text when refreshing the browser page, even with caching
disabled.

In this better implementation, the view calls the locale file and sends
the text via the data attribute to the JS, which works perfectly.

I also bumped the wait time for that one spec that fails intermittently
on Travis