Refactor pagination and make it customizable
The Kaminari gem has been around for a while and provides all the
functionality we need via clean and concise code, and allows us to
customize various aspects of the pagination.

Notable changes:

- The search results summary is now simple and only displays
information about the quantity of results. It does not mention any
search parameters, such as “matching ‘food’ within 5 miles of Redwood
City.” The reasoning behind that is that having conditional text based
on search parameters that might or not be there (some people might not
want to have certain search filters) can result in unused code.
Furthermore, the more parameters you have, the more complex the method
becomes, and the more space you need to display all that text. I’d
prefer the base code to be as simple and maintainable as possible, and
can be built upon as needed by people forking the project.

- The specs have been simplified to test the basics. There’s no need to
test how the numbers are displayed as you navigate through the pages
since that is handled by the Kaminari gem, which is already tested.

- The pagination now includes a link to the First and Last pages, and
those links need to be styled.

Customization:

- You can customize the text/html for all the navigation links in
`config/locales/en.yml`.

- In that same locales file, you can also customize the text of the
search results summary. `%{entry_name}` refers to what you want to call
the results. For example, if you want to say “Displaying 1 - 29 of 50
locations”, you can set `entry_name` to ‘location’. Note that
`entry_name` should be in the singular form. Currently, it is not
possible to customize `entry_name` from the locales, but it is a
feature being worked on in Kaminari and should be available soon. In
the meantime, you can set it in `app/helpers/results_summary_helper.rb`
and in
`app/views/component/organizations/results/_header_summary.html.haml`.

- You can customize the way page numbers are displayed by adjusting
various settings in `config/initializers/kaminari_config.rb`.

- You can customize the HTML for each pagination element in the various
partials in `app/views/kaminari`.

Issues:

- I couldn’t figure out how to get rid of the space between the page
numbers. In our previous implementation, the borders touched each
other, but now I can’t get them to touch. The HTML looks the same to
me, but maybe I overlooked something.