Allow Travis to run from external pull requests.
By default, the `bin/wad` script sets an Authorization header using S3
credentials for both read and write operations. The problem is that pull
requests from forks can't use secure credentials on Travis, so all
Travis
runs from external PRs will fail.

What gets stored on S3 is just a zipped up tarball of all the gems,
so it's perfectly safe to allow anyone to download the files from S3.
This commit modifies the script so that it only sets the Authorization
header during a write operation. On the S3 side, a policy was added to
the bucket to allow anonymous downloads:
http://docs.aws.amazon.com/AmazonS3/latest/dev/example-bucket-policies.h
tml