Use git strategy for deployment to staging.
By default, Travis uses Anvil to deploy, which runs the Heroku buildpack compilation on Anvil. This fails because the required env vars aren't defined in Travis, and they're not necessary to run the tests. By using the git strategy, the buildpack is run on a Heroku dyno (presumably the one for staging) which should already have the necessary env vars.