Clean up and improve asset precompilation.
When precompiling assets locally for production, the font-awesome .otf
file was missing, and a blank square was displayed instead of the font
when running the app in development. Replacing the font-awesome-sass
gem with font-awesome-rails fixed the bug: precompiling assets after
adding the gem added the otf file to the assets directory.

I also removed the precompilation line in application.rb since it was a
duplicate of what’s in assets.rb.

Finally, I replaced the sentry gem with app-signal for analytics.
AppSignal doesn’t require an initializer file that points to an API key
via an env var.