Recompile assets and bump version number.
For some reason, I kept getting 404s for all the map markers. Bumping
the assets version in assets.rb and precompiling assets fixed the issue.

Added `kind` to map JSON in search.rb so that results-map-manager can
use it.

Added spec to check for presence of libraries marker on map. It’s
failing because the JS needs to be fixed.