Cache API responses.
To improve the performance of Ohana Web Search, and to reduce the load on the API, we tell Ohana Web Search to cache API responses.

This is handled by the Ohanakapa Ruby wrapper in conjuction with the `faraday-http-cache` gem, which determines whether or not to repeat an API request based on the Cache-Control header in the API response.

The amount of time an API response is cached is determined by the API. By default, it is set to 5 minutes, a reasonable amount of time given that the API responses don't change that often.