Removes modernize selectors test script
For browsers that don’t support the CSS selector :checked for radio
buttons (e.g. IE8), the modernizer plugin modernizr.selectors was used
to add a test for the presence of this selector. The test for :checked
selector support was removed in
https://github.com/codeforamerica/ohana-web-search/commit/fd75f5d427ee94
84c0dc7e6fb751045e6b4ec8fa. This commit removes associated JavaScript
scripts for the selector test. If :checked is added to the CSS in the
future, this and
https://github.com/codeforamerica/ohana-web-search/commit/fd75f5d427ee94
84c0dc7e6fb751045e6b4ec8fa should be reversed.