Add HTTP caching to locations_controller
The server now includes a `Last-Modified` header, which the browser uses to ask the server if things have changed (via the `If-Modified-Since` Request header).

If the `Last-Modified` date has not changed, when a browser requests the same search results or location details page, the server will respond with a `304 Not Modified` and will skip rendering the view, thereby improving performance.

Other changes:

 - Organize and document Caching and Email delivery setups in production.rb

- Document password protection in staging.rb

- Remove logging from ohanapi.rb