Add service_area checkbox
When checked, the checkbox will restrict search results to those whose service areas field contains SMC service areas.

This is the simplest implementation since checking the box will add the service_area=smc parameter, and unchecking it will not submit the checkbox.

Note that farmers' markets don't show up currently when the filter is set because they don't have services. This is a data issue that can be fixed by updating them in the admin interface.

I also added the filter to all homepage links to be consistent with the search box.