Refactors Google Translate code
- Moves Google Translate initialization to application.js so it’s
applied to all routes.
- Refactors Google Translate Manager to remove polling code and to
integrate Google supplied initialization scripts.
- Refactors language translation link settings:
- Strips whitespace around language code setting so any amount of
spacing can be introduced into the settings and the app won’t silently
not configure the language code URL properly.
- Adds ability to remove all language links and the app won’t break.
Ties into #483.
- Edits language links code comments to be consistent with other
settings in the settings file and adds a space after the colon in the
settings to be consistent with Ruby style guide key/value styling.
- Refactors area above and below search box on homepage so that when
language links are absent the height of the utility area doesn’t
collapse. Also, renames bottom utility area to be more generic.
- Adds translation services section to customization readme.
- Moves google translate manager JS to `util` directory.
- Moves drop-down layout handling code to its own module.
- Capitalizes modules that include creation method for creating a new
instance.
- Adds error handling so if Google script changes container HTML an
error will be thrown.
- Fixes #287 Google Translate cookie wasn’t being set on inside pages.
This commit adds the Google Translate cookie concern to inside pages in
addition to the homepage.