Inverts appearance of service area toggle
Inverts the appearance of the service area toggle by providing a SASS
mixin that allows checkboxes to invert their appearance (checked looks
like unchecked and vice versa) based on whether an ‘inverted’ class is
present on the filter. It would be preferable to include the ‘inverted’
on the toggle-container, not on the filter fieldset, but for some
reason when the ‘active’ class was added the SASS/CSS wouldn’t
recognize that the inverted class was applied any more. A future
refactor could look into this, though it is not critical.