Fix Keyword Translator issue.
When you initialize the client, you need to pass in the application
name and version. Also, translations sometimes is an empty array, which
will pass the `nil?` test, so we should use `blank?` instead.

Also, note that if you don’t get any results for certain non-English
words, it could be because we’ve reached the daily limit, which
apparently we already have.