Update cassettes to match API and fix mailto spec.
After merging master into the branch, a couple specs were failing
because they were using the old `san-maceo-agency` location instead of
the new `example-location`. Deleting the cassettes and regenerating
them fixed the issue.

Because master added the mailto feature, and because this branch added
the alternate_name to the name, the mailto spec failed because the
mailto helper method was only using the base name whereas the spec was
grabbing the name from the h1.name element, which also includes the
alternate name. The fix was to add the alternate name to the mailto
helper.

By defining a `full_name_for(location)` method, we can now use it
instead of having logic in the views for formatting the location name.