Fixes #663 - adds phone helper for results
Fixes #663 - adds phone helper for search results so that the first
voice or hotline phone number is displayed on the search results list,
or none if either no number or only numbers of type `tty` or `fax` are
found. Also, singularizes HTML class from `phones` to `phone` for
search result entries and removes unneeded list for phones.