Fixes #667 - adds new service fields exposed API
- In order to organize the location details spec better, this commit
moves the location services tests to their own spec file.
- Orders service tests to fit the order shown in the detail view (with
the exception of the hours, which will be handled separately in
https://github.com/codeforamerica/ohana-web-search/issues/671)
- Updates service_areas service partial name to be in the format field
name prefixed by `service_`, which follows the format of the other
partials. The file was named `service_areas`, which implies the field
is `areas`, when in fact, it is `service_areas`.
- Removes optional requirement from `how_to_apply` service field.
- Edits `name` and adds `alternate_name` service field and tests.
Removes optional check for the required `name` field.
- Adds test for service description field
and updates specificity of description test.
- Makes location description test check the description selector contents
as opposed to the whole page, since the checked content “Lorem Ipsum”
appears in the service description fields as well.
- Adds `email` service field and test.
- Renames `urls` service field to `website`, since a service now only has
one website. Also adds a test.
- Adds a container and styles for the parent element around `website` and
`email` service fields.
-Adds `accepted_payments` service field, locale, and test.
-Adds `required_documents` service field, locale, and test.
- Adds `status` service field and test.
- Renames name partial container and CSS from `title` to `name-box`.